requirements.txt
# Core dependencies
python-dotenv==1.0.0
openai==1.12.0

# Email handling
imaplib2==3.6
secure-smtplib==0.1.1

# Document processing
PyPDF2==3.0.1
docx2txt==0.8

# Utility packages
regex==2023.12.25
python-dateutil==2.8.2

# Development dependencies
pytest==7.4.3
black==23.12.1
flake8==7.0.0 